requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0
sqlalchemy==2.0.29
pydantic==2.7.0
pydantic-settings==2.2.1
requests==2.31.0
python-multipart==0.0.9
psutil==5.9.8
python-dotenv==1.0.1
aiofiles==23.2.1
jinja2==3.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
waitress==3.0.0
alembic==1.13.1
google-generativeai==0.5.2
slowapi==0.1.9
# Soporte para Windows (Ignorado en Render/Linux)
pypiwin32==223; sys_platform == 'win32'
